created module and did some filename changes for better understanding with help of rahul's guidance
