+++ .env
@@ -1,1 +1,2 @@
-PASSWORD = ulsolqvvokxgdmiv
+PASSWORD=ulsolqvvokxgdmiv
+EMAIL=harishpanjikar@gmail.com
